backend/pyproject.toml
# Tuxedo Backend Dependencies
# Use with: uv add fastapi uvicorn pydantic python-dotenv langchain-openai stellar-sdk aiohttp certifi

[project]
name = "tuxedo-backend"
version = "0.1.0"
description = "FastAPI backend for Tuxedo AI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "langchain>=1.0.3",
    "langchain-openai>=0.1.0",
    "langchain-core>=1.0.1",
    "stellar-sdk[aiohttp]>=13.1.0",
    "aiohttp>=3.8.0",
    "certifi>=2022.12.7",
    "sendgrid>=6.12.5",
    "email-validator>=2.3.0",
    "webauthn>=2.0.0",
    "cryptography>=42.0.0",
    "aiodns>=3.5.0",
    "openhands-aci>=0.3.2",
    "openhands-tools>=1.2.0",
    "boto3>=1.41.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "openhands-sdk>=1.2.0",
    "openhands-workspace>=1.3.0",
    "openhands-agent-server>=1.3.0",
    "deepagents>=0.2.8",
    "langgraph>=1.0.1",
    "pre-commit>=4.5.0",
]
